refactor: introduce bulletproof
